refactored enemy manager and health component

+++ resources/EnemyResource.cs
@@ -9,9 +9,9 @@ public partial class EnemyResource : Resource
 	[Export] public string Name { get; set; }
 	[Export] public string Description { get; set; }
 	
-	[Export] public int TotalHealth { get; set; }
-	[Export] public int Damage { get; set; }
-	[Export] public int Speed { get; set; }
+	[Export] public float TotalHealth { get; set; }
+	[Export] public float Damage { get; set; }
+	[Export] public float Speed { get; set; }
 
 	[ExportGroup("Spawning Settings")]
 	[Export] public int SpawnChance { get; set; }
